requirements.txt
attrs==25.1.0
Automat==24.8.1
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
constantly==23.10.4
cryptography==44.0.1
cssselect==1.2.0
defusedxml==0.7.1
filelock==3.17.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.11.0
itemloaders==1.3.2
jmespath==1.0.1
lxml==5.3.1
packaging==24.2
parsel==1.10.0
Protego==0.4.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyDispatcher==2.0.7
pyOpenSSL==25.0.0
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
Scrapy==2.12.0
service-identity==24.2.0
soupsieve==2.6
tldextract==5.1.3
Twisted==24.11.0
typing_extensions==4.12.2
urllib3==2.3.0
w3lib==2.3.1
zope.interface==7.2
